% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqmeans.R
\name{seqmeans}
\alias{seqmeans}
\title{Executes a sequential algorithm for sample size calculations for longitudinal surveys}
\usage{
seqmeans(
  theta = 0.5,
  rho,
  deff,
  S2x,
  S2y,
  alt,
  del,
  sig.level = 0.05,
  power = 0.8,
  S2o,
  alt.gross,
  del.gross,
  sig.level.gross = 0.05,
  pow.gross = 0.8
)
}
\arguments{
\item{theta}{- description .............}

\item{rho}{- unit-level correlation between x and y}

\item{deff}{- description .............}

\item{S2x}{- unit variance of analysis variable x in sample t}

\item{S2y}{- unit variance of analysis variable y in sample t+1}

\item{alt}{- should the test of the net change be "1-sided" or "2-sided"}

\item{del}{- size of the net change between the means to be detected}

\item{sig.level}{- significance level of the hypothesis test of the net change}

\item{power}{- desired power of the test of the net change}

\item{S2o}{- description .............}

\item{alt.gross}{- should the test of the gross change be "1-sided" or "2-sided"}

\item{del.gross}{- size of the gross change between the means to be detected}

\item{sig.level.gross}{- significance level of the hypothesis test of the gross change}

\item{pow.gross}{- desired power of the test of the gross change}
}
\value{
panel sample size and overall sample size
}
\description{
\code{seqmeans} executes a sequential algorithm for sample size calculation to measure net and gross changes concurrently in longitudinal surveys.
\code{seqmeans} deals with changes in means
}
\examples{
# Calculate sample size of panel and fresh samples to estimate newt and gross changes

sample_sequence <- seqmeans(
theta = 0.5,
rho = 0.9,
deff = 1,
S2x= 100,
S2y= 75,
alt= "one.sided",
del=3,
sig.level = 0.05,
power = 0.80,
S2o=200,
alt.gross="one.sided",
del.gross=3,
sig.level.gross = 0.05,
pow.gross = 0.80)

}
\author{
Mahmoud Elkasabi.
}
