% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqprop.R
\name{seqprop}
\alias{seqprop}
\title{Executes a sequential algorithm for sample size calculations for longitudinal surveys}
\usage{
seqprop(
  theta = 0.5,
  deff,
  P1,
  P2,
  PXY,
  alt,
  sig.level = 0.05,
  power = 0.8,
  P1.gross,
  P2.gross,
  PXY.gross,
  alt.gross,
  sig.level.gross = 0.05,
  pow.gross = 0.8
)
}
\arguments{
\item{theta}{- description .............}

\item{deff}{- description .............}

\item{P1}{- proportion in survey t}

\item{P2}{- proportion in survey t+1}

\item{PXY}{- proportion in the overlap has the characteristic in both samples}

\item{alt}{- should the test of the net change be "1-sided" or "2-sided"}

\item{sig.level}{- significance level of the hypothesis test of the net change}

\item{power}{- desired power of the test of the net change}

\item{P1.gross}{- proportion in survey t among panel sample}

\item{P2.gross}{- proportion in survey t+1 among panel sample}

\item{PXY.gross}{- proportion in the overlap has the characteristic in both panel samples}

\item{alt.gross}{- should the test of the gross change be "1-sided" or "2-sided"}

\item{sig.level.gross}{- significance level of the hypothesis test of the gross change}

\item{pow.gross}{- desired power of the test of the gross change}

\item{del}{- size of the net change between the means to be detected}

\item{del.gross}{- size of the gross change between the means to be detected}
}
\value{
panel sample size and overall sample size
}
\description{
\code{seqprop} executes a sequential algorithm for sample size calculation to measure net and gross changes concurrently in longitudinal surveys.
\code{seqprop} deals with changes in proportions
}
\examples{
# Calculate sample size of panel and fresh samples to estimate newt and gross changes

sample_sequence <- seqprop(
theta = 0.5,
deff = 1,
P1= 0.5,
P2= 0.7,
PXY= 0.45,
alt= "one.sided",
sig.level = 0.05,
power = 0.80,
P1.gross= 0.5,
P2.gross= 0.75,
PXY.gross= 0.45,
alt.gross="one.sided",
sig.level.gross = 0.05,
pow.gross = 0.80)

}
\author{
Mahmoud Elkasabi.
}
